requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
streamlit
python-dotenv
ollama
tiktoken
rapidfuzz
